+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Tests for "Graffiti on the fence".
 - Tests for "Prefix code".
 - Tests for "Offset Arrays".
+- Tests for "Unit fractions".
 
 ## [3.7.0] - 2022-03-31
 ### Added